Revert "Revert "Revert "P20 1575 fix check for account upgrades"" (#69560)"
This reverts commit e37378e237685c436dc4ad4b18363a473072748a.
