+++ dashboard/app/controllers/registrations_controller.rb
@@ -388,8 +388,6 @@ class RegistrationsController < Devise::RegistrationsController
     return head(:bad_request) if params[:user].nil?
     return head(:bad_request) if params[:user][:user_type].nil?
 
-    previous_user_type = current_user.user_type
-
     successfully_updated =
       if current_user.migrated?
         if forbidden_change?(current_user, params)
@@ -412,18 +410,6 @@ class RegistrationsController < Devise::RegistrationsController
           current_user.update_without_password(set_user_type_params)
         end
       end
-
-    if successfully_updated && (previous_user_type != current_user.user_type)
-      Metrics::Events.log_event(
-        user: current_user,
-        event_name: 'user_type_changed',
-        metadata: {
-          from_user_type: previous_user_type,
-          to_user_type: current_user.user_type
-        },
-        session: session
-      )
-    end
 
     if successfully_updated
       head :no_content
