+++ dashboard/app/helpers/users_helper.rb
@@ -267,7 +267,7 @@ module UsersHelper
     # Therefore, we don't need to show the cap_user_info modal.
     return false if !Policies::ChildAccount.user_predates_state_collection?(user) && user.us_state.present?
     # Is the student a child and using a personal account to access code.org?
-    user.under_13? && Policies::User.personal_account?(user)
+    user.under_13? && Policies::ChildAccount.personal_account?(user)
   end
 
   def lti_user_info_required?(user)
